api_gateway/requirements.txt
# Development/Testing Dependencies
pytest>=7.0.0
pytest-mock>=3.0.0
requests>=2.25.0 # For sending requests in tests

# Core Dependencies
fastapi
uvicorn[standard]
grpcio
grpcio-tools
protobuf
neo4j

grpcio-tools>=1.48.0 # For generating code during build